+++ tests/core/test_integration.py
@@ -19,6 +19,8 @@ from pytest_mock.plugin import MockerFixture
 from sqlglot import exp
 from sqlglot.expressions import DataType
 import re
+from IPython.utils.capture import capture_output
+
 
 from sqlmesh import CustomMaterialization
 from sqlmesh.cli.example_project import init_example_project
@@ -62,10 +64,11 @@ from sqlmesh.core.snapshot import (
     SnapshotTableInfo,
 )
 from sqlmesh.utils.date import TimeLike, now, to_date, to_datetime, to_timestamp
-from sqlmesh.utils.errors import NoChangesPlanError, SQLMeshError
+from sqlmesh.utils.errors import NoChangesPlanError, SQLMeshError, PlanError
 from sqlmesh.utils.pydantic import validate_string
 from tests.conftest import DuckDBMetadata, SushiDataValidator
 from tests.utils.test_helpers import use_terminal_console
+from tests.utils.test_filesystem import create_temp_file
 
 if t.TYPE_CHECKING:
     from sqlmesh import QueryOrDF
@@ -6085,3 +6088,42 @@ def test_destroy(copy_to_temp_path):
 
     # Ensure the cache has been removed
     assert not cache_path.exists()
+
+
+@use_terminal_console
+def test_audits_running_on_metadata_changes(tmp_path: Path):
+    def setup_senario(model_before: str, model_after: str):
+        models_dir = Path("models")
+        create_temp_file(tmp_path, models_dir / "test.sql", model_before)
+
+        # Create first snapshot
+        context = Context(paths=tmp_path, config=Config())
+        context.plan("prod", no_prompts=True, auto_apply=True)
+
+        # Create second (metadata) snapshot
+        create_temp_file(tmp_path, models_dir / "test.sql", model_after)
+        context.load()
+
+        with capture_output() as output:
+            with pytest.raises(PlanError):
+                context.plan("prod", no_prompts=True, auto_apply=True)
+
+        assert 'Failed models\n\n  "model"' in output.stdout
+
+        return output
+
+    # Ensure incorrect audits (bad data, incorrect definition etc) are evaluated immediately
+    output = setup_senario(
+        "MODEL (name model); SELECT NULL AS col",
+        "MODEL (name model, audits (not_null(columns=[col]))); SELECT NULL AS col",
+    )
+    assert "'not_null' audit error: 1 row failed" in output.stdout
+
+    output = setup_senario(
+        "MODEL (name model); SELECT NULL AS col",
+        "MODEL (name model, audits (not_null(columns=[this_col_does_not_exist]))); SELECT NULL AS col",
+    )
+    assert (
+        'Binder Error: Referenced column "this_col_does_not_exist" not found in \nFROM clause!'
+        in output.stdout
+    )
